Fixes issue with repository log view
The addressee display template was updated to include a check for PII, but the
implementation assumes that the item being managed is an article in a
journal, but that template fragment is also used by the repository log.

Adds some logic to check if the article variable is defined before
applying PII checks. This means that PII checks are not appliedd to
repository email addresses in this list, but as none of the rest of the
PII functionality is built out for repository side, this seems
reasonble.

Fixes #4637